Main Menu:
On App Open, the main menu will show. The main menu has three different buttons, when they are hovered over, the GUI on the right changes based on the topic shown by the button's text--for example, the credits button shows the credits and sources when its hovered over.

There is no click functionality in the buttons yet (the only one with functionality will be the "Begin Plan" button, of which it will take the user to the next scene, where it will ask them to select a region to place their firm).

Animations will be added later, if not next. Still trying to figure out a theme, design, and/or color scheme for the application.