Regions
I added images for the next scenes as well as buttons (that dont work just yet) to take the user to the next scene with the map of the region they selected.

I plan to then ask the user where they want their firm inside of the region of which they will place a pin anywhere they want with a set radius and then firms selected from a list of coordinates will appear.

The user will then go to the next scene to reveal their score calculated by the firms that were in the local radius of the firm they placed.