Region Scenes
I got images of each of the seven regions (NA, CA, SA, EU, AF, AS, OC), and made their own scenes out of them. I added click functionality to the buttons and I also added a nice (can be improved but I don't have the will) fade animation to compliment the transition between scenes for user experience points.

I added elements to each of the scenes, which can definitely be refactored to not only erase redundant assets but also code & allow it to be more readable.

I added functionality for each of the region scenes' back buttons to take the user back to the region selection map if they wish to change their mind.

I plan to add a scene between the main menu and the region selection scene called the firm selection scene where the user picks the type of firm they wish to make.

The region scene would then update the back button to go back to the firm selection instead of the main menu if the user wishes to change their mind on what type of firm they wish to publish.

I plan to add a fourth button in the main menu to take the user to the logs, the logs is a place where they can see their last firm placements (it will save upon exit & load upon opening the app) IF AND ONLY IF I HAVE TIME.

I plan to add a .txt file to be parsed into the directory where each entry will look something like this:

"REGION_CoordX_CoordY_TYPE"

for example:
"CA_104_108_SC" (semiconductor)

there will be a second file to maybe be parsed that is a glossary that translates the entry's respective elements.

for example:

"AS_390_278_RM"

Region: Asia
Coords: 390, 278
Type of firm: RAW MATERIALS

This is a rough roadmap I am setting for myself, gonna have to work my butt off to implement the systems, please do help if you have an idea.

& please do help fill out the list of different firms, there is a lot to go through.

the coordinate system is up for discussion, but one I am currently exampling in this commit is one where it simply matches the pixel dimensions of the image it is on.

I will create a doc or an obsidian note that completely explains everything, I will allocate times to spend time on each part of the project, or you can take roles in explaining the app.