Update On Region Scenes
I updated each image to all be the same dimensions for better coordinate system compatibility & tracking compatibility

one feature it allowed me to add is a simple cursor of a pin and a circle that follows the mouse whenever the mouse is hovering over the region map.

Next thing on the agenda for the scene is when the user clicks, it will disable the event method (that has the pin follow the mouse) and places the pin (likely accompanied by an animation) and save the circle position, and then run a loop to gather every firm's coordinates and create small circles of differing colors based on what type of firm it is at their location, and then it will highlight the firm if it lies within the radius circle the user placed, of which it is counted toward the score.