Refactoring Step 1
I added a "Helpers" class to store the helper methods that were previously stored in Main (of which was renamed from HelloApplication).

I plan on refactoring a bunch of code and adding it to a third file where it sets up all of the apps scenes & buttons with their own instance variables, etc.