Redundancy & Bug Squashing + Front-end development
I scrapped the firm picking process of the app.

I got rid of redundant methods that are not going to be used, I'll soon get rid of elements that aren't used.

I fixed some bugs mainly with button hitboxes.

With adjusting button hitboxes I had to remake the alignment of the buttons by adding a padding style ENUM to add more customizability with the recently refactored createBackButtonLayout.