Penultimate commit for the app:
I implemented Aeden's scoring system
We all got 30 samples of different firm locations of different resource types
When you place the pin, you are guided to the score results where it tells you the region you selected, the type you were to look for, and the score you got. the score goes from 100 to 10,000, the closer to 10,000 the closer your pin placement was to the firm location that was collected in the sample. the scoring system goes by distance to the closest possible firm.

I made it possible to do as many tries as the user wants by adding a reset() method that resets the app and takes the user to the home menu for them to go at it again.

The only thing left to do is ensure resolution changes have no bugs, documentation, an explanation pdf of the app, and to fill the other regions' lists of firm locations